move test_pack_with_mobile_version to other file
